% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_MNI_hmm.R
\name{build_T_from_dist}
\alias{build_T_from_dist}
\title{Construir matriz de transición T (un paso de dt horas)}
\usage{
build_T_from_dist(dist_mat, alpha)
}
\arguments{
\item{dist_mat}{distance matrix}

\item{alpha}{numeric}
}
\description{
Construir matriz de transición T (un paso de dt horas)
}
\keyword{internal}
