% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_independ.R
\name{event_independ}
\alias{event_independ}
\title{Identify Independent Camera-Trap Events}
\usage{
event_independ(
  input = NULL,
  station_field = NULL,
  species_field = NULL,
  individ_field = NULL,
  date_field = NULL,
  date_format = "\%d/\%m/\%Y",
  time_field = NULL,
  tz = "UTC",
  independent_time = 1,
  independent_units = "hours"
)
}
\arguments{
\item{input}{\code{data.frame}. A table containing the camera-trap records.}

\item{station_field}{\code{character}. Name of the column in \code{input} that uniquely
identifies each camera station. If set to \code{NULL}, all records are assumed to
come from a single station.}

\item{species_field}{\code{character}. Name of the column containing the species
identifier. If \code{NULL}, all records are assumed to belong to the same species.}

\item{date_field}{\code{character}. Name of the column containing the \strong{date} of
each record. If set to \code{NULL}, all records are assumed to belong to the same day.}

\item{date_format}{\code{character}. Date format used in the records. The default is
\code{"\%d/\%m/\%Y"}, where \verb{\%d} = day, \verb{\%m} = month, and \verb{\%Y} = year, separated by \code{"/"}.
If your dates use a different structure, adjust accordingly.
Examples:
\code{"02/23/2023"} → \code{"\%m/\%d/\%Y"}
\code{"02-23-2023"} → \code{"\%m-\%d-\%Y"}}

\item{time_field}{\code{character}. Name of the column containing the \strong{time} of
each record. The required format is always \code{"\%H:\%M:\%S"} (Hour:Minute:Second).}

\item{tz}{\code{character}. Time zone used for date–time parsing. The default is \code{"UTC"}
(recommended), but other time zones may be specified depending on the country
or region of the study.}

\item{independent_time}{\code{numeric}. Minimum time required between records to
consider them independent events. The value is interpreted according to the unit
specified in \code{independent_units}. For example, if \code{independent_units = "weeks"}
and \code{independent_time = 1}, then two records are considered independent only if
at least one week has passed. Default is \code{1}.}

\item{independent_units}{\code{character}. Time unit used to define independence
between events. Options are \code{"secs"}, \code{"mins"}, \code{"hours"}, \code{"days"}, or \code{"weeks"}.
Default is \code{"hours"}.}

\item{indiv_field}{\code{character}. Name of the column indicating the number of
individuals recorded in each photograph. See Example 3 for details.}
}
\description{
\code{event_independ()} processes camera-trap detection data and identifies
independent events based on a user-defined temporal threshold. The function
allows independence to be evaluated by station, species, and (optionally)
number of individuals in the photo. Records occurring within the specified
time window are grouped together, and only one record is retained per group,
either the first occurrence or the one with the highest number of individuals
if an \code{indiv_field} is provided.
}
\details{
This function is useful for standardizing camera-trap datasets by reducing
temporal autocorrelation, estimating relative abundance indices, preparing
data for activity pattern analyses, and ensuring consistency across studies
that define independent detections using temporal rules (e.g., 30 minutes,
1 hour, 24 hours). The function accommodates multiple stations, species,
time zones, and various date–time formats.
}
\examples{
\dontrun{
datab <- read.csv("base_barranca.csv", header = TRUE)

#Example 1. Use a 1-hour threshold to define independent events for each
#species and station.

Example1 <- event_independ(input = datab,
                           station_field = "Station",
                           species_field = "Species",
                           date_field = "Date",
                           date_format = "\%d/\%m/\%Y",
                           time_field = "Time",
                           independent_time = 1,
                           independent_units = "hours")
Example1

#Example 2. Use a 2-day threshold to define independent events for each
#species and station.

Example2 <- event_independ(input = datab,
                           station_field = "Station",
                           species_field = "Species",
                           date_field = "Date",
                           date_format = "\%d/\%m/\%Y",
                           time_field = "Time",
                           independent_time = 2,
                           independent_units = "days")
Example2

#Example 3. Use a 1-hour threshold to define independent events for each
#species and station, giving priority to records with more individuals.
#For instance, if three records (a, b, c) fall within a one-hour interval
#and the number of individuals is a = 1, b = 3, and c = 2, then the first
#record (a) is retained when indiv_field = NULL.
#If a field for the number of individuals is provided and indiv_field is not NULL,
#then record b is retained because it has the highest number of individuals.

#Example with fake data: add a new column containing random values of 2 and 3 individuals.
datab$Individuos <- 1
datab$Individuos[sample(1:nrow(datab), 100)] <- sample(c(2, 3), 100, replace = TRUE)

Example3 <- event_independ(input = datab,
                           station_field = "Station",
                           species_field = "Species",
                           indiv_field = "Individuos",
                           date_field = "Date",
                           date_format = "\%d/\%m/\%Y",
                           time_field = "Time",
                           independent_time = 1,
                           independent_units = "hours")
Example3
}
}
