% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hor_min_sec_rad.R
\name{hor_min_sec_rad}
\alias{hor_min_sec_rad}
\title{Convert time (HH:MM:SS) to radians}
\usage{
hor_min_sec_rad(x)
}
\arguments{
\item{x}{character. vector with HH:MM:SS}
}
\description{
Convert time (HH:MM:SS) to radians
}
