% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_iNEXT.R
\name{input_iNEXT_cameras}
\alias{input_iNEXT_cameras}
\title{input_iNEXT_cameras}
\usage{
input_iNEXT_cameras(
  input = NULL,
  species_field = "Species",
  assemblages = NULL,
  station_field = "Station",
  type = c("raw", "frequency")
)
}
\arguments{
\item{input}{data.frame}

\item{species_field}{character}

\item{assemblages}{character}

\item{station_field}{character}

\item{type}{character}
}
\description{
input_iNEXT_cameras
}
\keyword{internal}
