% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_iNEXT.R
\name{input_iNEXT_days}
\alias{input_iNEXT_days}
\title{input_iNEXT_days}
\usage{
input_iNEXT_days(
  input = NULL,
  species_field = "Species",
  assemblages = NULL,
  date_field = "Date",
  date_format = "\%d/\%m/\%Y",
  type = c("raw", "frequency")
)
}
\arguments{
\item{input}{data.frame}

\item{species_field}{character}

\item{assemblages}{character}

\item{date_field}{character}

\item{date_format}{character}

\item{type}{character}
}
\description{
input_iNEXT_days
}
\keyword{internal}
