% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_effort.R
\name{sampling_effort}
\alias{sampling_effort}
\title{Sampling effort}
\usage{
sampling_effort(
  input = NULL,
  method = 1,
  station_field = "Station",
  season_field = NULL,
  date_field = "Date",
  date_format = "\%d/\%m/\%Y"
)
}
\arguments{
\item{input}{\code{data.frame}. (\emph{required}) Table containing the camera-trap data.}

\item{method}{\code{integer}. (\emph{required}) Method used to estimate sampling effort.
Two options are available:

\itemize{
\item \strong{\code{1}}: Does not discriminate between the minimum and maximum date
of each individual camera. The input table is split by year. For each year,
the earliest (first photo) and latest (last photo) dates across all cameras
are computed. The number of days between these two dates is multiplied by
the number of cameras. This process is repeated for each year, and
camera-days/nights are summed across years.

\if{html}{\out{<div class="sourceCode">}}\preformatted{\\itemize\{
  \\item If `season_field` is not `NULL`, each year is further subdivided
  by sampling periods (e.g., dry vs. wet season), preventing the inclusion
  of inactive periods.
  \\item This method may **overestimate** sampling effort, because it
  assumes all cameras operated for the same number of days.
\}
}\if{html}{\out{</div>}}

\item \strong{\code{2}}: Similar to method 1, but the calculation is performed
\strong{per camera}. For each camera, the number of days between its first and
last photo is computed, then results are summed across cameras.

\if{html}{\out{<div class="sourceCode">}}\preformatted{\\itemize\{
  \\item This method may **underestimate** sampling effort, as it depends
  entirely on the records actually captured by each camera.
\}
}\if{html}{\out{</div>}}

}}

\item{station_field}{\code{character}. Name of the column in the \code{data.frame} that
contains the unique identifier for each station. If \code{NULL}, all records are
assumed to belong to a single station.}

\item{season_field}{\code{character}. Name of the column representing any category
used to distinguish among different sampling periods (e.g., wet vs. dry season).
If \code{NULL}, sampling is assumed to be continuous and all records belong to a
single period.}

\item{date_field}{\code{character}. Name of the column containing the date of each
record. If \code{NULL}, all records are assumed to occur on the same day.}

\item{date_format}{\code{character}. Date format used in the data. The default is
\code{"\%d/\%m/\%Y"}, where \verb{\%d} = day, \verb{\%m} = month, and \verb{\%Y} = year, separated by
\code{"/"}. Change the format according to your dataset.

Examples:
\code{"02/23/2023"} → \code{"\%m/\%d/\%Y"}
\code{"02-23-2023"} → \code{"\%m-\%d-\%Y"}}
}
\description{
Estimate Camera-Trap Sampling Effort
}
\details{
\code{effort_estimate()} calculates the sampling effort (e.g., camera-days or
camera-nights) from camera-trap detection data using two alternative methods.
The function relies on the minimum and maximum sampling dates recorded either
across all cameras (method 1) or for each camera individually (method 2).
Optionally, sampling periods such as wet and dry seasons can be considered to
avoid counting inactive intervals.

The function is designed to support standardized effort calculations in
ecological studies based on camera-trap data, enabling comparisons among
stations, seasons, and years. Depending on the chosen method, the resulting
effort may represent an upper- or lower-bound estimate, providing flexibility
for different study designs and data availability.
}
\examples{
\dontrun{
datab <- read.csv("sipecam_final.csv", header = TRUE)

samp_eff <- sampling_effort(input = datab,
                            method = 1,
                            station_field = "Station",
                            season_field = NULL,
                            date_field = "Date",
                            date_format = "\%d/\%m/\%Y")
samp_eff

samp_eff <- sampling_effort(input = datab,
                            method = 2,
                            station_field = "Station",
                            season_field = NULL,
                            date_field = "Date",
                            date_format = "\%d/\%m/\%Y")
samp_eff

samp_eff <- sampling_effort(input = datab,
                            method = 1,
                            station_field = "Station",
                            season_field = "Muestreo",
                            date_field = "Date",
                            date_format = "\%d/\%m/\%Y")
samp_eff

samp_eff <- sampling_effort(input = datab,
                            method = 2,
                            station_field = "Station",
                            season_field = "Muestreo",
                            date_field = "Date",
                            date_format = "\%d/\%m/\%Y")
samp_eff
}
}
