% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_meta_table.R
\name{video_meta_table}
\alias{video_meta_table}
\title{Video metadata to record table}
\usage{
video_meta_table(
  folderpath,
  tags = "TagsList",
  date_field = "MetadataDate",
  date_format = "\%Y:\%m:\%d \%H:\%M:\%S",
  ext = ".xmp"
)
}
\arguments{
\item{folderpath}{\code{character}. The path to the folder with the video metadata;
if there are subfolders inside, the search will include them.}

\item{tags}{\code{character}. The name of the column that holds video tags. It usually includes
the species name but might also contain other tags, such as the number of individuals or
a specific behavior.}

\item{date_field}{\code{character}. Name of the column containing the \strong{date and time} of
each record.}

\item{date_format}{\code{character}. Date format used in the records. The default is
\code{"\%d/\%m/\%Y"}, where \verb{\%d} = day, \verb{\%m} = month, and \verb{\%Y} = year, separated by \code{"/"}.
If your dates use a different structure, adjust accordingly.
Examples:
\code{"2026/02/17 01:29:23"} → \code{"\%Y/\%m/\%d \%H:\%M:\%S"}
\code{"02:23:2023 01:29:23"} → \code{"\%m-\%d-\%Y \%H:\%M:\%S"}}

\item{ext}{\code{character}. Metadata file extension, default = ".xmp" (see exifr::read_exif())}
}
\description{
Use the video metadata to create a table with data from the camera trap records,
including the date, time, and tags.
}
\examples{
\dontrun{
#Carpeta con subcarpetas
test <- video_meta_table(folderpath = "D:/Camaras_trampaR/Videos_prueba2",
                         tags = "TagsList",
                         date_field = "MetadataDate",
                         date_format = "\%Y:\%m:\%d \%H:\%M:\%S",
                         ext = ".xmp")
head(test)
}
}
